BCStheoryCriticalCurrentFit{32,64}.m: resolve unicode problems
We cannot use Tex in uicontrol, and char() is limited to 255 in
Octave. Instead just put the characters in the file and save the file
as UTF instead of ASCII. And make some other buttons a bit nicer while
changing this. Works for me now in both Octave and MatLab.

This resolves the following runtime warning:

warning: range error for conversion to character value

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>